+++ tests/test_healpix_padding.py
@@ -113,13 +113,13 @@ def test_healpix_pad_cuda_channels_last(nchannels, dtype):
     padding = 1
     n = 3
     x = torch.randn([n, ntile, nchannels, nside, nside], device="cuda", dtype=dtype)
-    x_channels_last = x.permute(0, 1, 3, 4, 2).contiguous()
+    x_channels_last = x.permute(0, 1, 3, 4, 2).contiguous().permute(0, 1, 4, 2, 3)
     x.requires_grad = True
     x_channels_last.requires_grad = True
     with pad_backend(PaddingBackends.cuda):
         out_channels_first = pad(x, padding=padding)
-        out_channels_last = pad(x_channels_last, padding=padding, channels_last=True)
+        out_channels_last = pad(x_channels_last, padding=padding)
     out_channels_first.mean().backward()
     out_channels_last.mean().backward()
-    assert torch.allclose(out_channels_first, out_channels_last.permute(0, 1, 4, 2, 3))
-    assert torch.allclose(x.grad, x_channels_last.grad.permute(0, 1, 4, 2, 3))
+    assert torch.allclose(out_channels_first, out_channels_last)
+    assert torch.allclose(x.grad, x_channels_last.grad)